Read extended description, many changes
Removed the bounding box defining where trackers can be placed. Changed the script to have a custom sized screenshot size soon to be defined by the user in the Settings GUI.
Everything is now relative to the custom size. The main player can actually be detected currently, but this will be fixed in the future.