Many improvements, see description
Added a new "ignore box" that should prevent detections of the own player by setting the confidence level of detections made in that area to 0.

Added error handling for a prominent error I was having with playing audio. Please run the "run_spectre.bat" file from now on to prevent this (hopefully!).